feat(bet4-release): extract release & promotion tools into a scoped ToolPack (EP-8DC217EB BET-4)
Process-Spine-Decision: BET-4 drain-to-zero (plan §4) — behavior-identical pack extraction; imports shared logBuildActivity from build-tool-helpers.
BI: BI-2B7EE073
