+++ apps/web/lib/mcp-tools.ts
@@ -1433,112 +1433,6 @@ export const PLATFORM_TOOLS: ToolDefinition[] = [
     executionMode: "immediate",
     sideEffect: true, // Tier 3: potentially destructive
   },
-  // ─── Scheduling & Release Tools (IT4IT §5.3-5.4) ───────────────────────
-  {
-    name: "check_deployment_windows",
-    description: "Check available deployment windows for promoting changes to production. Returns current window status, blackout periods, and next available window time.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        change_type: { type: "string", description: "RFC type: standard, normal, or emergency. Default: normal." },
-        risk_level: { type: "string", description: "Risk level: low, medium, high, or critical. Default: low." },
-      },
-    },
-    requiredCapability: "view_operations",
-    executionMode: "immediate",
-    sideEffect: false,
-    buildPhases: ["review", "ship"],
-  },
-  {
-    name: "schedule_promotion",
-    description: "Schedule an approved promotion for deployment during a specific window. Creates a calendar event for visibility.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        promotion_id: { type: "string", description: "The promotion ID (CP-xxx) to schedule." },
-      },
-      required: ["promotion_id"],
-    },
-    requiredCapability: "view_operations",
-    executionMode: "immediate",
-    sideEffect: true,
-    buildPhases: ["ship"],
-  },
-  {
-    name: "execute_promotion",
-    description: "Execute an approved promotion. Starts the autonomous promoter: backup DB, build new portal image from sandbox, swap containers, health check. Rolls back automatically on failure.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        promotion_id: { type: "string", description: "The promotion ID to execute (e.g. CP-xxxx)." },
-        override_reason: { type: "string", description: "Reason for deploying outside a deployment window (optional, for emergency changes)." },
-      },
-      required: ["promotion_id"],
-    },
-    requiredCapability: "view_operations" as const,
-    executionMode: "immediate" as const,
-    sideEffect: true,
-    buildPhases: ["ship"],
-  },
-  {
-    name: "create_release_bundle",
-    description: "Group multiple completed builds into a release bundle for coordinated deployment (IT4IT §5.3.5 Release Package).",
-    inputSchema: {
-      type: "object",
-      properties: {
-        title: { type: "string", description: "Release bundle title, e.g. 'March 2026 Feature Release'." },
-        build_ids: { type: "array", items: { type: "string" }, description: "Array of buildId values to include in the bundle." },
-      },
-      required: ["title", "build_ids"],
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: true,
-    buildPhases: ["ship"],
-  },
-  {
-    name: "run_release_gate",
-    description: "Run gate checks on a release bundle: combine diffs from all builds, run destructive operation scan, validate all builds passed tests (IT4IT §5.3.5 Accept & Publish Release).",
-    inputSchema: {
-      type: "object",
-      properties: {
-        bundle_id: { type: "string", description: "The release bundle ID (RB-xxx) to check." },
-      },
-      required: ["bundle_id"],
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: true,
-  },
-  {
-    name: "schedule_release_bundle",
-    description: "Schedule an approved release bundle for deployment during a deployment window. Creates an RFC, ChangePromotion, and CalendarEvent for operations calendar visibility.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        bundle_id: { type: "string", description: "The release bundle ID (RB-xxx) to schedule." },
-      },
-      required: ["bundle_id"],
-    },
-    requiredCapability: "view_operations",
-    executionMode: "immediate",
-    sideEffect: true,
-  },
-  {
-    name: "get_release_status",
-    description: "Get the current status of a release bundle or promotion, including deployment window availability and gate check results.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        bundle_id: { type: "string", description: "Release bundle ID (RB-xxx) — optional if promotion_id is provided." },
-        promotion_id: { type: "string", description: "Promotion ID (CP-xxx) — optional if bundle_id is provided." },
-      },
-    },
-    requiredCapability: "view_operations",
-    executionMode: "immediate",
-    sideEffect: false,
-    buildPhases: ["ship"],
-  },
   // ─── Hive Mind Contribution Tools (IT4IT §5.5 Release) ───────────────────
   {
     name: "assess_contribution",
@@ -5880,549 +5774,6 @@ export async function executeTool(
           gates: gateResult,
         },
       };
-    }
-
-    // ─── Scheduling & Release Tools ──────────────────────────────────────────
-
-    case "check_deployment_windows": {
-      const changeType = String(params.change_type ?? "normal");
-      const riskLevel = String(params.risk_level ?? "low");
-      const profile = await prisma.businessProfile.findFirst({
-        where: { isActive: true },
-        include: { deploymentWindows: true, blackoutPeriods: true },
-      });
-
-      if (!profile) {
-        return { success: true, message: "No business profile configured — deployment is unrestricted. Set up operating hours in Admin to enable deployment windows.", data: { available: true, unrestricted: true } };
-      }
-
-      const now = new Date();
-      const activeBlackout = profile.blackoutPeriods.find(
-        (bp) => bp.startAt <= now && bp.endAt >= now && !bp.exceptions.includes(changeType),
-      );
-      if (activeBlackout) {
-        return {
-          success: true,
-          message: `Blackout period active until ${activeBlackout.endAt.toISOString()}. Reason: ${activeBlackout.reason ?? "Scheduled blackout"}. Emergency changes may override.`,
-          data: { available: false, blackout: true, blackoutEnd: activeBlackout.endAt.toISOString(), reason: activeBlackout.reason },
-        };
-      }
-
-      const { isNowInWindow } = await import("@/lib/sandbox-promotion");
-      const matchingWindows = profile.deploymentWindows.filter(
-        (w) => w.allowedChangeTypes.includes(changeType) && w.allowedRiskLevels.includes(riskLevel),
-      );
-
-      if (matchingWindows.length === 0) {
-        return { success: true, message: "No deployment windows configured for this change type and risk level — deployment is unrestricted.", data: { available: true, unrestricted: true } };
-      }
-
-      const windowOpen = isNowInWindow(matchingWindows);
-      const windowSummary = matchingWindows.map((w) => ({
-        name: w.name,
-        days: w.dayOfWeek,
-        startTime: w.startTime,
-        endTime: w.endTime,
-      }));
-
-      return {
-        success: true,
-        message: windowOpen
-          ? `Deployment window is OPEN now. ${matchingWindows.length} matching window(s) available.`
-          : `Not in a deployment window. Available windows: ${matchingWindows.map((w) => `${w.name}: days ${w.dayOfWeek.join(",")}, ${w.startTime}-${w.endTime}`).join("; ")}`,
-        data: { available: windowOpen, windows: windowSummary },
-      };
-    }
-
-    case "schedule_promotion": {
-      const promotionId = String(params.promotion_id ?? "");
-      if (!promotionId) return { success: false, error: "promotion_id is required.", message: "Provide a promotion ID." };
-
-      const promotion = await prisma.changePromotion.findUnique({
-        where: { promotionId },
-        include: {
-          changeItem: { include: { changeRequest: true } },
-          productVersion: { include: { featureBuild: { select: { buildId: true } } } },
-        },
-      });
-      if (!promotion) return { success: false, error: "Promotion not found.", message: `No promotion with ID ${promotionId}.` };
-      if (promotion.status !== "approved") return { success: false, error: "Promotion must be approved first.", message: `Current status: ${promotion.status}` };
-
-      // Find next available window
-      const profile = await prisma.businessProfile.findFirst({
-        where: { isActive: true },
-        include: { deploymentWindows: true },
-      });
-
-      if (!profile || profile.deploymentWindows.length === 0) {
-        return { success: true, message: "No deployment windows configured. Promotion can be deployed anytime via Operations > Promotions.", data: { scheduled: false } };
-      }
-
-      const rfcType = promotion.changeItem?.changeRequest?.type ?? "normal";
-      const riskLevel = promotion.changeItem?.changeRequest?.riskLevel ?? "low";
-      const matchingWindows = profile.deploymentWindows.filter(
-        (w) => w.allowedChangeTypes.includes(rfcType) && w.allowedRiskLevels.includes(riskLevel),
-      );
-
-      if (matchingWindows.length === 0) {
-        return { success: true, message: "No windows match this change type and risk level. Ask an admin to configure appropriate deployment windows.", data: { scheduled: false } };
-      }
-
-      // Update RFC with deployment window info
-      const rfc = promotion.changeItem?.changeRequest;
-      if (rfc) {
-        await prisma.changeRequest.update({
-          where: { id: rfc.id },
-          data: {
-            status: "scheduled",
-            scheduledAt: new Date(),
-            deploymentWindowId: matchingWindows[0]!.id,
-          },
-        });
-      }
-
-      const windowDesc = matchingWindows.map((w) => `${w.name}: days ${w.dayOfWeek.join(",")}, ${w.startTime}-${w.endTime}`).join("; ");
-
-      // Persist the scheduled state on the ChangePromotion so getBuildFlowState
-      // reads it without peeking at ChangeRequest. deploymentLog carries the
-      // window description so the fork can render "Scheduled for <window>".
-      await prisma.changePromotion.update({
-        where: { promotionId },
-        data: { status: "scheduled", deploymentLog: `window: ${windowDesc}` },
-      }).catch(() => {});
-
-      logBuildActivity(promotionId, "schedule_promotion", `Scheduled for window: ${windowDesc}`);
-
-      const scheduleBuildId = promotion.productVersion?.featureBuild?.buildId ?? null;
-      if (scheduleBuildId) {
-        const { reconcileBuildCompletion } = await import("@/lib/build-flow-state");
-        await reconcileBuildCompletion(scheduleBuildId).catch(() => {});
-      }
-
-      return {
-        success: true,
-        message: `Promotion ${promotionId} scheduled. Deployment windows: ${windowDesc}. An operator can deploy via Operations > Promotions during an open window.`,
-        data: { scheduled: true, windows: windowDesc },
-      };
-    }
-
-    case "execute_promotion": {
-      const promotionId = String(params.promotion_id ?? "");
-      const overrideReason = params.override_reason ? String(params.override_reason) : undefined;
-      if (!promotionId || !/^[a-zA-Z0-9_-]+$/.test(promotionId)) {
-        return { success: false, error: "Invalid promotion_id", message: "Provide a valid promotion ID." };
-      }
-
-      // Validate promotion exists and is approved
-      const promo = await prisma.changePromotion.findFirst({ where: { promotionId } });
-      if (!promo) return { success: false, error: "Not found", message: `Promotion ${promotionId} not found.` };
-      if (promo.status === "deployed") return { success: true, message: "Already deployed.", data: { status: "deployed" } };
-      if (promo.status !== "approved") return { success: false, error: `Status is ${promo.status}`, message: "Must be approved first." };
-
-      // Enforce deployment window — block execution outside windows unless emergency override
-      if (!overrideReason) {
-        const { getPromotionWindowStatus } = await import("@/lib/actions/promotions");
-        const windowStatus = await getPromotionWindowStatus(promotionId);
-        if (!windowStatus.available) {
-          return {
-            success: false,
-            error: "Outside deployment window",
-            message: `${windowStatus.message} Use schedule_promotion to queue for the next window, or provide override_reason for emergency deployment.`,
-          };
-        }
-      }
-
-      // Resolve sandbox and build ID
-      const promoDetail = await prisma.changePromotion.findFirst({
-        where: { promotionId },
-        include: { productVersion: { include: { featureBuild: { select: { sandboxId: true, buildId: true } } } } },
-      });
-      const sandboxId = promoDetail?.productVersion?.featureBuild?.sandboxId;
-      const promoBuildId = promoDetail?.productVersion?.featureBuild?.buildId;
-      if (!sandboxId) return { success: false, error: "No sandbox", message: "No sandbox linked to this promotion." };
-
-      const { execFile: execFileCb } = lazyChildProcess();
-      const { promisify } = lazyUtil();
-      const execFileAsync = promisify(execFileCb);
-      const execAsync = promisify((lazyChildProcess()).exec);
-
-      // Preflight: the promoter image has to exist locally for `docker run
-      // dpf-promoter` to work. On most installs today it doesn't — the image
-      // is built separately and isn't on the default compose path. Detect
-      // that up front and hand off to the operator UI instead of attempting
-      // the run and returning a scary "Could not start the promoter
-      // container." message. The promotion stays approved; the operator
-      // triggers it from Operations > Promotions.
-      try {
-        await execAsync("docker image inspect dpf-promoter");
-      } catch {
-        // Persist the awaiting_operator state so getBuildFlowState reads it
-        // from the ChangePromotion column directly (no BuildActivity detour)
-        // and the reconciler can treat the fork as dispositioned. The prior
-        // version returned awaiting_operator only in the tool response, which
-        // the fork state machine couldn't see.
-        await prisma.changePromotion.update({
-          where: { promotionId },
-          data: { status: "awaiting_operator" },
-        }).catch(() => {});
-        logBuildActivity(promoBuildId ?? promotionId, "execute_promotion", "Promoter image not present — handing off to operator");
-        if (promoBuildId) {
-          const { reconcileBuildCompletion } = await import("@/lib/build-flow-state");
-          await reconcileBuildCompletion(promoBuildId).catch(() => {});
-        }
-        return {
-          success: true,
-          message: `Promotion ${promotionId} is approved and ready. This install doesn't have the "dpf-promoter" container image built locally, so automatic deployment is disabled here. An operator needs to run the promotion manually from Operations > Promotions, which will stream the deployment log and handle rollback. The promotion stays in "awaiting_operator" status until then.`,
-          data: { status: "awaiting_operator", reason: "promoter_image_missing", promotionId },
-        };
-      }
-
-      // Start promoter container (array form — no shell injection)
-      try {
-        await execAsync("docker rm dpf-promoter-1 2>/dev/null || true");
-        const dockerArgs = [
-          "run", "-d",
-          "--name", "dpf-promoter-1",
-          "--network", `${process.env.DPF_COMPOSE_PROJECT ?? "dpf"}_default`,
-          "-v", "/var/run/docker.sock:/var/run/docker.sock",
-          "-v", "dpf_backups:/backups",
-          "-e", `PROMOTION_ID=${promotionId}`,
-          "-e", `DPF_PRODUCTION_DB_CONTAINER=${process.env.DPF_PRODUCTION_DB_CONTAINER ?? "dpf-postgres-1"}`,
-          "-e", "DPF_PORTAL_CONTAINER=dpf-portal-1",
-          "-e", `DPF_COMPOSE_PROJECT=${process.env.DPF_COMPOSE_PROJECT ?? "dpf"}`,
-          "-e", `DPF_SANDBOX_CONTAINER=${sandboxId}`,
-          "-e", `POSTGRES_USER=${process.env.POSTGRES_USER ?? "dpf"}`,
-        ];
-        if (overrideReason) {
-          dockerArgs.push("-e", `DPF_WINDOW_OVERRIDE=${overrideReason}`);
-        }
-        dockerArgs.push("dpf-promoter");
-        await execFileAsync("docker", dockerArgs);
-      } catch (err) {
-        return {
-          success: false,
-          error: `Failed to start promoter: ${(err as Error).message?.slice(0, 200)}`,
-          message: `Could not start the promoter container. An operator can run this promotion manually from Operations > Promotions — the promotion stays in "approved" status until deployed.`,
-        };
-      }
-
-      // Poll for completion (max 10 minutes)
-      const maxWaitMs = 10 * 60 * 1000;
-      const pollIntervalMs = 10_000;
-      const startTime = Date.now();
-      let exitCode: number | null = null;
-
-      while (Date.now() - startTime < maxWaitMs) {
-        await new Promise(r => setTimeout(r, pollIntervalMs));
-        try {
-          const { stdout } = await execAsync("docker inspect dpf-promoter-1 --format='{{.State.Status}} {{.State.ExitCode}}'");
-          const parts = stdout.trim().replace(/'/g, "").split(" ");
-          if (parts[0] === "exited") {
-            exitCode = parseInt(parts[1] ?? "1", 10);
-            break;
-          }
-        } catch { /* container may not exist yet */ }
-      }
-
-      if (exitCode === null) {
-        await execAsync("docker stop dpf-promoter-1 2>/dev/null || true").catch(() => {});
-        return { success: false, error: "Timeout (10 min)", message: "Promoter did not complete. Check ops dashboard." };
-      }
-
-      const finalPromo = await prisma.changePromotion.findFirst({ where: { promotionId } });
-      const promoSuccess = exitCode === 0 && finalPromo?.status === "deployed";
-
-      await execAsync("docker rm dpf-promoter-1 2>/dev/null || true").catch(() => {});
-      logBuildActivity(promoBuildId ?? promotionId, "execute_promotion", promoSuccess ? "Deployed successfully" : `Rolled back: ${finalPromo?.rollbackReason ?? "unknown"}`);
-
-      if (promoBuildId) {
-        const { reconcileBuildCompletion } = await import("@/lib/build-flow-state");
-        await reconcileBuildCompletion(promoBuildId).catch(() => {});
-      }
-
-      return {
-        success: promoSuccess,
-        message: promoSuccess
-          ? `Promotion ${promotionId} deployed. Health check passed.`
-          : `Rolled back. ${finalPromo?.rollbackReason ?? "Check deployment log."}`,
-        data: { promotionId, status: finalPromo?.status, deploymentLog: finalPromo?.deploymentLog?.slice(0, 1000) },
-      };
-    }
-
-    case "create_release_bundle": {
-      const title = String(params.title ?? "");
-      const buildIds = Array.isArray(params.build_ids) ? params.build_ids.map(String) : [];
-      if (!title) return { success: false, error: "title is required.", message: "Provide a release bundle title." };
-      if (buildIds.length === 0) return { success: false, error: "build_ids is required.", message: "Provide at least one build ID." };
-
-      // Validate all builds exist and are in review/complete phase
-      const builds = await prisma.featureBuild.findMany({
-        where: { buildId: { in: buildIds } },
-        select: { buildId: true, title: true, phase: true, releaseBundleId: true },
-      });
-      const missing = buildIds.filter((id) => !builds.some((b) => b.buildId === id));
-      if (missing.length > 0) return { success: false, error: `Builds not found: ${missing.join(", ")}`, message: `Could not find builds: ${missing.join(", ")}` };
-
-      const notReady = builds.filter((b) => !["review", "complete", "ship"].includes(b.phase));
-      if (notReady.length > 0) {
-        return { success: false, error: `Builds not ready: ${notReady.map((b) => `${b.buildId} (${b.phase})`).join(", ")}`, message: `All builds must be in review or complete phase.` };
-      }
-
-      const alreadyBundled = builds.filter((b) => b.releaseBundleId);
-      if (alreadyBundled.length > 0) {
-        return { success: false, error: `Builds already in a bundle: ${alreadyBundled.map((b) => b.buildId).join(", ")}`, message: `Remove from existing bundle first.` };
-      }
-
-      // Create the bundle
-      const bundleId = `RB-${new Date().toISOString().slice(0, 10)}-${crypto.randomUUID().slice(0, 6).toUpperCase()}`;
-      const bundle = await prisma.releaseBundle.create({
-        data: {
-          bundleId,
-          title,
-          status: "assembling",
-          createdBy: userId,
-        },
-      });
-
-      // Link builds to the bundle
-      await prisma.featureBuild.updateMany({
-        where: { buildId: { in: buildIds } },
-        data: { releaseBundleId: bundle.id },
-      });
-
-      return {
-        success: true,
-        message: `Release bundle ${bundleId} created with ${buildIds.length} build(s): ${builds.map((b) => b.title).join(", ")}. Run gate checks before scheduling deployment.`,
-        data: { bundleId, title, buildCount: buildIds.length, builds: builds.map((b) => ({ buildId: b.buildId, title: b.title })) },
-      };
-    }
-
-    case "run_release_gate": {
-      const bundleId = String(params.bundle_id ?? "");
-      if (!bundleId) return { success: false, error: "bundle_id is required.", message: "Provide a release bundle ID." };
-
-      const bundle = await prisma.releaseBundle.findUnique({
-        where: { bundleId },
-        include: {
-          builds: {
-            select: {
-              buildId: true, title: true, phase: true, diffPatch: true,
-              verificationOut: true, sandboxId: true,
-            },
-          },
-        },
-      });
-      if (!bundle) return { success: false, error: "Bundle not found.", message: `No release bundle with ID ${bundleId}.` };
-      if (bundle.status !== "assembling") {
-        return { success: false, error: `Gate check already run. Bundle status: ${bundle.status}`, message: `Bundle is in ${bundle.status} state.` };
-      }
-      if (bundle.builds.length === 0) {
-        return { success: false, error: "Bundle has no builds.", message: "Add builds to the bundle first." };
-      }
-
-      // Check all builds are in review/complete/ship phase
-      const notReady = bundle.builds.filter((b) => !["review", "complete", "ship"].includes(b.phase));
-      if (notReady.length > 0) {
-        return {
-          success: false, error: `Builds not ready: ${notReady.map((b) => `${b.buildId} (${b.phase})`).join(", ")}`,
-          message: "All builds must be in review or complete phase.",
-        };
-      }
-
-      // Check all builds have passing tests
-      const failedTests = bundle.builds.filter((b) => {
-        const v = b.verificationOut as Record<string, unknown> | null;
-        return v && (v.testsPassed === false || v.testsPassed === 0);
-      });
-
-      // Combine diffs from all builds
-      const diffs: string[] = [];
-      for (const build of bundle.builds) {
-        if (build.diffPatch) {
-          diffs.push(build.diffPatch as string);
-        } else if (build.sandboxId) {
-          try {
-            const { extractDiff } = await import("@/lib/sandbox");
-            const diff = await extractDiff(build.sandboxId);
-            diffs.push(diff);
-          } catch {
-            // Build has no extractable diff — may be fine if it's code-only
-          }
-        }
-      }
-
-      const combinedDiff = diffs.join("\n");
-
-      // Scan for destructive operations
-      const { scanForDestructiveOps, categorizeDiffFiles } = await import("@/lib/sandbox-promotion");
-      const allFileMatches = [...combinedDiff.matchAll(/^diff --git a\/(.+) b\/.+$/gm)].map((m) => m[1]);
-      const { migrationFiles } = categorizeDiffFiles(allFileMatches);
-      const destructiveWarnings = migrationFiles.length > 0 ? scanForDestructiveOps(combinedDiff) : [];
-
-      // Build gate check result
-      const gateResult = {
-        buildsChecked: bundle.builds.length,
-        allTestsPass: failedTests.length === 0,
-        failedTestBuilds: failedTests.map((b) => b.buildId),
-        totalFilesChanged: allFileMatches.length,
-        migrationFiles: migrationFiles.length,
-        destructiveWarnings,
-        combinedDiffLength: combinedDiff.length,
-      };
-
-      const passed = failedTests.length === 0 && destructiveWarnings.length === 0;
-
-      // Update bundle
-      await prisma.releaseBundle.update({
-        where: { bundleId },
-        data: {
-          status: passed ? "approved" : "gate_check",
-          combinedDiffPatch: combinedDiff,
-          gateCheckResult: gateResult as unknown as import("@dpf/db").Prisma.InputJsonValue,
-        },
-      });
-
-      const messageParts = [
-        `Gate check ${passed ? "PASSED" : "FAILED"} for ${bundleId}.`,
-        `${bundle.builds.length} build(s), ${allFileMatches.length} file(s) changed, ${migrationFiles.length} migration(s).`,
-      ];
-      if (failedTests.length > 0) messageParts.push(`Failing tests in: ${failedTests.map((b) => b.buildId).join(", ")}.`);
-      if (destructiveWarnings.length > 0) messageParts.push(`Destructive ops: ${destructiveWarnings.join("; ")}`);
-      if (passed) messageParts.push("Bundle is approved and ready to schedule for deployment.");
-
-      return { success: true, message: messageParts.join(" "), data: gateResult };
-    }
-
-    case "schedule_release_bundle": {
-      const bundleId = String(params.bundle_id ?? "");
-      if (!bundleId) return { success: false, error: "bundle_id is required.", message: "Provide a release bundle ID." };
-
-      const bundle = await prisma.releaseBundle.findUnique({
-        where: { bundleId },
-        include: { builds: { select: { buildId: true, title: true, createdById: true } } },
-      });
-      if (!bundle) return { success: false, error: "Bundle not found.", message: `No bundle ${bundleId}.` };
-      if (bundle.status !== "approved") {
-        return { success: false, error: `Bundle must be approved first. Current: ${bundle.status}`, message: `Run gate checks first.` };
-      }
-
-      // Find next available deployment window
-      const profile = await prisma.businessProfile.findFirst({
-        where: { isActive: true },
-        include: { deploymentWindows: true },
-      });
-
-      const matchingWindows = profile?.deploymentWindows.filter(
-        (w) => w.allowedChangeTypes.includes("normal") && w.allowedRiskLevels.includes("low"),
-      ) ?? [];
-
-      // Create RFC for the bundle
-      const rfcId = `RFC-${new Date().getFullYear()}-${crypto.randomUUID().slice(0, 8).toUpperCase()}`;
-      const rfc = await prisma.changeRequest.create({
-        data: {
-          rfcId,
-          title: `Release: ${bundle.title}`,
-          description: `Release bundle ${bundleId} with ${bundle.builds.length} build(s): ${bundle.builds.map((b) => b.title).join(", ")}`,
-          type: "normal",
-          scope: "platform",
-          riskLevel: "low",
-          status: "scheduled",
-          scheduledAt: new Date(),
-          requestedById: bundle.createdBy,
-          ...(matchingWindows.length > 0 ? { deploymentWindowId: matchingWindows[0]!.id } : {}),
-        },
-      });
-
-      // Create CalendarEvent for visibility
-      const employee = await prisma.employeeProfile.findFirst({
-        where: { userId: bundle.createdBy },
-        select: { id: true },
-      });
-
-      let calendarEventId: string | undefined;
-      if (employee) {
-        const eventId = `RELEASE-${bundleId}`;
-        await prisma.calendarEvent.upsert({
-          where: { eventId },
-          create: {
-            eventId,
-            title: `Deployment: ${bundle.title}`,
-            description: `${bundle.builds.length} feature(s): ${bundle.builds.map((b) => b.title).join(", ")}`,
-            startAt: new Date(),
-            eventType: "action",
-            category: "platform",
-            ownerEmployeeId: employee.id,
-            visibility: "team",
-            color: "#f59e0b",
-          },
-          update: { title: `Deployment: ${bundle.title}`, startAt: new Date() },
-        });
-        calendarEventId = eventId;
-      }
-
-      // Update bundle
-      await prisma.releaseBundle.update({
-        where: { bundleId },
-        data: {
-          status: "scheduled",
-          rfcId: rfc.rfcId,
-          calendarEventId,
-          scheduledAt: new Date(),
-          ...(matchingWindows.length > 0 ? { deploymentWindowId: matchingWindows[0]!.id } : {}),
-        },
-      });
-
-      const windowDesc = matchingWindows.length > 0
-        ? matchingWindows.map((w) => `${w.name}: days ${w.dayOfWeek.join(",")}, ${w.startTime}-${w.endTime}`).join("; ")
-        : "No windows configured — deployment unrestricted";
-
-      return {
-        success: true,
-        message: `Release ${bundleId} scheduled. RFC: ${rfcId}. Added to operations calendar. Windows: ${windowDesc}. An operator can deploy via Operations > Promotions.`,
-        data: { bundleId, rfcId, calendarEventId, windows: windowDesc },
-      };
-    }
-
-    case "get_release_status": {
-      const bundleId = params.bundle_id ? String(params.bundle_id) : null;
-      const promotionId = params.promotion_id ? String(params.promotion_id) : null;
-
-      if (bundleId) {
-        const bundle = await prisma.releaseBundle.findUnique({
-          where: { bundleId },
-          include: { builds: { select: { buildId: true, title: true, phase: true } } },
-        });
-        if (!bundle) return { success: false, error: "Bundle not found.", message: `No release bundle with ID ${bundleId}.` };
-        return {
-          success: true,
-          message: `Release ${bundle.bundleId}: ${bundle.status}. ${bundle.builds.length} build(s).`,
-          data: {
-            bundleId: bundle.bundleId,
-            title: bundle.title,
-            status: bundle.status,
-            builds: bundle.builds,
-            scheduledAt: bundle.scheduledAt?.toISOString() ?? null,
-            deployedAt: bundle.deployedAt?.toISOString() ?? null,
-          },
-        };
-      }
-
-      if (promotionId) {
-        const { getPromotionWindowStatus } = await import("@/lib/actions/promotions");
-        const windowStatus = await getPromotionWindowStatus(promotionId).catch(() => ({ available: false, message: "Could not check window status" }));
-        const promotion = await prisma.changePromotion.findUnique({
-          where: { promotionId },
-          select: { status: true, deployedAt: true, rationale: true, rollbackReason: true },
-        });
-        if (!promotion) return { success: false, error: "Promotion not found.", message: `No promotion with ID ${promotionId}.` };
-        return {
-          success: true,
-          message: `Promotion ${promotionId}: ${promotion.status}. Window: ${windowStatus.message}`,
-          data: { promotionId, ...promotion, windowStatus },
-        };
-      }
-
-      return { success: false, error: "Provide bundle_id or promotion_id.", message: "Specify which release or promotion to check." };
     }
 
     // ─── Hive Mind Contribution ──────────────────────────────────────────────
