+++ scripts/mcp-tool-pack-baseline.json
@@ -9,19 +9,16 @@
   "approve_decomposition",
   "assess_complexity",
   "assess_contribution",
-  "check_deployment_windows",
   "check_sandbox",
   "contribute_to_hive",
   "create_build_epic",
   "create_digital_product",
   "create_portal_pr",
-  "create_release_bundle",
   "deploy_feature",
   "diagnose_sandbox",
   "dpf_test_kernel_refuse_probe",
   "drive_browser_task",
   "edit_sandbox_file",
-  "execute_promotion",
   "generate_code",
   "get_build_dispatch_history",
   "get_build_engine_readiness",
@@ -29,7 +26,6 @@
   "get_build_sandbox_state",
   "get_build_scoped_verification",
   "get_finance_period_summary",
-  "get_release_status",
   "inspect_build_code_impact",
   "iterate_sandbox",
   "list_build_activity_since",
@@ -56,7 +52,6 @@
   "reviewBuildPlan",
   "reviewDesignDoc",
   "run_endpoint_tests",
-  "run_release_gate",
   "run_sandbox_command",
   "run_sandbox_tests",
   "run_tool_script",
@@ -64,8 +59,6 @@
   "saveBuildEvidence",
   "save_build_notes",
   "save_phase_handoff",
-  "schedule_promotion",
-  "schedule_release_bundle",
   "search_sandbox",
   "set_change_disposition",
   "setup_email",
